19w36a

+++ net/minecraft/world/entity/monster/MagmaCube.java
@@ -118,8 +118,8 @@ extends Slime {
     }
 
     @Override
-    protected int getAttackDamage() {
-        return super.getAttackDamage() + 2;
+    protected float getAttackDamage() {
+        return super.getAttackDamage() + 2.0f;
     }
 
     @Override
